requirements.txt
odfpy
wordcloud
matplotlib
